app/requirements.txt
bring-api
fastapi
mashumaro
orjson
python-dotenv
uvicorn[standard]